requirements.txt
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.4
bokeh==2.3.3
cachetools==4.2.4
certifi==2020.6.20
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fontawesomefree==6.1.1
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.18
graphviz==0.19.1
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.5.1
notebook==6.4.10
numpy==1.18.5
openpyxl==3.0.10
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
pyarrow==6.0.1
pycparser==2.21
pydeck==0.6.2
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywaffle==1.1.0
pywin32==304
pywinpty==1.1.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
rich==12.4.4
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.0
semver==2.13.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
smart-open==6.0.0
smmap==5.0.0
streamlit==1.10.0
streamlit-aggrid==0.2.2.post2
streamlit-option-menu==0.3.2
streamlit-plotly-events==0.0.6
tenacity==8.0.1
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
vega==3.6.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
wincertstore==0.2
xlrd==2.0.1
zipp==3.6.0
